+++ src/AzzyBot-Next/Services/Modules/AzuraCastApiService.cs
@@ -554,6 +554,35 @@ public sealed class AzuraCastApiService(ILogger<AzuraCastApiService> logger, DbA
         AzuraStationStatusRecord status = await GetFromApiAsync<AzuraStationStatusRecord>(baseUrl, endpoint, CreateHeader(apiKey));
         if (!status.BackendRunning || !status.FrontendRunning)
             throw new InvalidOperationException("Station failed to start.");
+
+        if (!config.BackendConfig.WritePlaylistsToLiquidsoap)
+            return;
+
+        // Delay to ensure the station is fully started
+        await Task.Delay(TimeSpan.FromSeconds(5));
+        AzuraNowPlayingDataRecord? nowPlaying = null;
+        try
+        {
+            nowPlaying = await GetNowPlayingAsync(baseUrl, stationId);
+        }
+        catch (HttpRequestException)
+        {
+            await context.EditResponseAsync("You have activated the option **\"Always Write Playlists to Liquidsoap\"** which means you have to wait more time until you can finally use your station.\nI inform you when it's finished.");
+        }
+
+        while (nowPlaying is null)
+        {
+            try
+            {
+                nowPlaying = await GetNowPlayingAsync(baseUrl, stationId);
+            }
+            catch (HttpRequestException)
+            {
+                await Task.Delay(TimeSpan.FromSeconds(10));
+            }
+        }
+
+        await context.FollowupAsync("All playlists have been written back to liquidsoap.");
     }
 
     public async Task StopStationAsync(Uri baseUrl, string apiKey, int stationId)